release candidate
updated the generator with command line args.  updated double qoutes to single for consistency (I might alter this later)

also updated the exporter so that it renames files that are too long

+++ a8_generator.py
@@ -2,8 +2,9 @@ import os, shutil, glob, re, argparse, sys
 
 ## Using the parser for command line arguments
 parser = argparse.ArgumentParser()
-parser.add_argument("-p", "--preset", required=True, help="name of the preset")
-parser.add_argument("-d", "--directory", required=True, help="location of the samples")
+parser.add_argument('-p', '--preset', required=True, help='This is the name of the preset')
+parser.add_argument('-n', '--number', required=True, help='This is the number of the preset, which is set inside the yml')
+parser.add_argument('-d', '--directory', required=True, help='location of the samples')
 #Not yet implmented
 parser.add_argument("-v", "--verbose", help="prints the yml configuration to stdout", action="store_true")
 args = vars(parser.parse_args())
@@ -12,28 +13,28 @@ args = vars(parser.parse_args())
 vt_nerdseq_001 = ['+4.56', '+4.48', '+4.39', '+4.31', '+4.23', '+4.14', '+4.04', '-5.00']
 
 ## This is the name of the preset which works nicely if it's the same name as your sample source folder
-preset = args["preset"]
+preset = args['preset']
 
 ## First make sure the path to the samples is valid
-if os.path.isdir(args["directory"]):
-    sample_dir = args["directory"] + '/' + preset
+if os.path.isdir(args['directory']):
+    sample_dir = args['directory'] + '/' + preset
 
 ## String building once for later use
-dest_dir = args["directory"] + '/completed_' + preset
+dest_dir = args["directory"] + '/completed_'# + preset
 
 ## Check if the destination directory exists, if it does, delete it and create it agian.  If not, create it!
 if not os.path.exists(dest_dir):
     os.mkdir(dest_dir)
-else:
-    shutil.rmtree(dest_dir)
-    os.mkdir(dest_dir)
+#else:
+#    shutil.rmtree(dest_dir)
+#    os.mkdir(dest_dir)
 
 ## Build a list of the channel directories that are not empty, and print the beginnings of the yml
 channels = []
 file_path = dest_dir + '/' + preset + '.yml'
 #print(file_path)
 with open(file_path, 'w') as f:
-    f.write('Preset 1 :\r\n')
+    f.write('Preset ' + args['number'] + ':\r\n')
     #print('Preset 1 :')
     f.write('  Name : ' + sample_dir[len(sample_dir)-7:len(sample_dir)]+'\r\n')
     #print('  Name : ' + sample_dir[len(sample_dir)-7:len(sample_dir)])
@@ -64,10 +65,7 @@ with open(file_path, 'w') as f:
                 for file in filenames:
                     (shortname, extension) = os.path.splitext(file)
                     if extension == '.wav':
-                        if len(shortname) > 47:
-                            samples.append(shortname[:-(len(shortname)-47)] + extension)
-                        else:
-                            samples.append(file)
+                        samples.append(file)
 
             samples.sort()
             j = 0
@@ -90,5 +88,5 @@ with open(file_path, 'w') as f:
 for root, dirnames, filenames in os.walk(sample_dir):
     for file in filenames:
         (shortname, extension) = os.path.splitext(file)
-        if extension == '.wav' and os.path.join(root,file) != sample_dir:
+        if extension == '.wav' and os.path.join(root,file) != sample_dir and not os.path.exists(os.path.join(dest_dir, file)):
             shutil.copy2(os.path.join(root,file), os.path.join(dest_dir,file))
